requirements.txt
sympy==1.8
matplotlib==3.4.2
python_telegram_bot==13.5
networkx==2.5.1
numpy==1.19.4
